+++ RobotSDK/ModuleDev/wheelchair/Sensor/stm32comm/StorageMono/Edit/StorageMono_Sensor_stm32comm_PrivFunc.cpp
@@ -89,7 +89,7 @@ bool DECOFUNC(processMonoDrainData)(void * paramsPtr, void * varsPtr, QVector<vo
 	Function: process draindata.
 	*/
     SourceDrainMono_Sensor_stm32comm_Data *data = draindata.front();
-    long time = ((data->timestamp.hour()*60 + data->timestamp.minute())*60  + data->timestamp.second()) *1000+data->timestamp.msec();
+    int time = ((data->timestamp.hour()*60 + data->timestamp.minute())*60  + data->timestamp.second()) *1000+data->timestamp.msec();
 
     vars->storagefile<<time<<data->yaw<<'\t'<<data->leftodom<<'\t'<<data->rightodom<<'\t'<<data->x<<'\t'<<data->y<<'\t'<<data->theta<<'\t'<<std::endl;
 	return 1;
